requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
requests==2.31.0
langchain==0.0.340
langchain-community==0.0.3
ollama==0.1.7
faiss-cpu==1.7.4
sentence-transformers==2.2.2
crewai==0.1.15
crewai_tools==0.1.6
pandas==2.1.3
numpy==1.24.4
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
logging==0.4.9.6
asyncio==3.4.3
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0